Refactor UIKit polluters
