+++ motion/ui/pollute.rb
@@ -1,5 +1,8 @@
+# Please, no more!  It'll hurt BubbleWrap's compatibility with other libraries.
 [
-  [UIControl, UIControlWrap]
+  [UIControl,         BW::UIControlWrapper],
+  [UIView,            BW::UIViewWrapper],
+  [UIViewController,  BW::UIViewControllerWrapper],
 ].each do |base, wrapper|
     base.send(:include, wrapper)
 end
